requirements.txt
bertopic
sentence-transformers
umap-learn
hdbscan
scikit-learn
matplotlib
seaborn
pandas
openpyxl
spacy
openai
nltk